event cache: introduce a `RoomPagination` API object and move code around
Only code motion. No changes in functionality.
