+++ crates/matrix-sdk-ui/src/timeline/pagination.rs
@@ -72,9 +72,11 @@ impl super::Timeline {
         let initial_options = options.clone();
         let mut outcome = PaginationOutcome::default();
 
+        let pagination = self.event_cache.pagination();
+
         while let Some(batch_size) = options.next_event_limit(outcome) {
             loop {
-                let result = self.event_cache.paginate_backwards(batch_size).await;
+                let result = pagination.paginate_backwards(batch_size).await;
 
                 let event_cache_outcome = match result {
                     Ok(outcome) => outcome,
@@ -145,7 +147,7 @@ impl super::Timeline {
     /// Note: this may send multiple Paginating/Idle sequences during a single
     /// call to [`Self::paginate_backwards()`].
     pub fn back_pagination_status(&self) -> Subscriber<PaginatorState> {
-        self.event_cache.pagination_status()
+        self.event_cache.pagination().pagination_status()
     }
 }
